pyproject.toml
[tool.poetry]
name = "dictdatabase"
version = "2.4.6"
repository = "https://github.com/mkrd/DictDataBase"
description = "Easy-to-use database using dicts"
authors = ["Marcel Kröker <kroeker.marcel@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers=[
	"Programming Language :: Python :: 3",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
	"Intended Audience :: Developers",
	"Programming Language :: Python",
	"Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.8"
orjson = "^3.8.7"

[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.10.0"
super-py = "^0.4.2"
pyinstrument = "^4.4.0"
pytest-cov = "^4.0.0"
path-dict = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
show-fixes = true
line-length = 120
select = [
	"ANN", # annotations
    "B", # bugbear
    "C", # comprehensions
    "E", # style errors
    "F", # flakes
    "I", # import sorting
    "M", # meta
    "N", # naming
    "U", # upgrade
    "W", # style warnings
    "YTT", # sys.version
]
ignore = [
	"ANN101", # missing type annotation for self
	"E501", # line length
	"UP007", # use X | Y for union (not possible in python 3.8)
    "UP006", # Use typing.Tuple for python 3.8 support
    "W191", # indentation contains tabs
    "E741", # ambiguous variable name
]

[tool.ruff.format]
indent-style = "tab"
quote-style = "double"
